requirements.txt
Django>=4.2
djangorestframework
djoser
drf-yasg
